requirements.txt
asgiref==3.5.0
attrs==23.2.0
Django==4.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.16.2
sqlparse==0.4.2
uritemplate==4.1.1
